+++ MovieApp/MovieApp/Views/ChatList/Views/CreateChatButton/CreateChatButton.swift
@@ -1,0 +1,20 @@
+//
+//  CreateChatButton.swift
+//  MovieApp
+//
+//  Created by Bobr, Andrey on 25.09.23.
+//
+
+import UIKit
+
+class CreateChatButton: UIButton {
+
+    /*
+    // Only override draw() if you perform custom drawing.
+    // An empty implementation adversely affects performance during animation.
+    override func draw(_ rect: CGRect) {
+        // Drawing code
+    }
+    */
+
+}
